优化SHT3X参数类型

+++ components/sht3x/include/sht3x.h
@@ -1,3 +1,3 @@
 esp_err_t sht3x_mode_init(void);
-uint8_t sht3x_get_humiture_periodic(uint16_t *Tem_val,uint16_t *Hum_val);
+uint8_t sht3x_get_humiture_periodic(float *Tem_val,float *Hum_val);
 uint8_t SHT3x_ReadSerialNumber(uint32_t* serialNumber);
